requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lz==0.11.1
MarkupSafe==3.0.2
memoir==0.0.3
mypy==0.991
mypy_extensions==1.1.0
mysqlclient==2.2.7
paradigm==0.6.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reprit==0.9.0
six==1.17.0
son==0.4.2
typing_extensions==4.13.2
Werkzeug==3.1.3

# Test dependencies
pytest==8.0.2
pytest-cov==4.1.0
mongomock==4.1.2
selenium==4.18.1
webdriver_manager==4.0.1
